Remove unfinished write handlers and move read handlers.
