+++ mr/src/main/java/org/elasticsearch/hadoop/serialization/ScrollReader.java
@@ -41,9 +41,9 @@ import org.elasticsearch.hadoop.serialization.builder.ValueReader;
 import org.elasticsearch.hadoop.serialization.dto.mapping.Mapping;
 import org.elasticsearch.hadoop.serialization.field.FieldFilter;
 import org.elasticsearch.hadoop.serialization.field.FieldFilter.NumberedInclude;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationErrorHandler;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationFailure;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationHandlerLoader;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationErrorHandler;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationFailure;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationHandlerLoader;
 import org.elasticsearch.hadoop.serialization.handler.SerdeErrorCollector;
 import org.elasticsearch.hadoop.serialization.json.BlockAwareJsonParser;
 import org.elasticsearch.hadoop.serialization.json.JacksonJsonParser;
