--- mr/src/main/java/org/elasticsearch/hadoop/serialization/handler/SerializationErrorHandler.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.hadoop.serialization.handler;
-
-import java.util.Properties;
-
-import org.elasticsearch.hadoop.handler.ErrorCollector;
-import org.elasticsearch.hadoop.handler.ErrorHandler;
-import org.elasticsearch.hadoop.handler.HandlerResult;
-
-/**
- * {@link ErrorHandler} subclass that allows for the interception of failures related to serializing integration specific
- * record types into JSON data to be written to Elasticsearch.
- * @param <R> The type of the integration specific record.
- */
-public abstract class SerializationErrorHandler<R> implements ErrorHandler<SerializationFailure, R, ErrorCollector<R>> {
-    public void init(Properties properties) {}
-
-    public abstract HandlerResult onError(SerializationFailure entry, ErrorCollector<R> collector) throws Exception;
-
-    public void close() {}
-}
