+++ mr/src/test/java/org/elasticsearch/hadoop/serialization/ScrollReaderTest.java
@@ -36,9 +36,9 @@ import org.elasticsearch.hadoop.serialization.ScrollReader.ScrollReaderConfig;
 import org.elasticsearch.hadoop.serialization.builder.JdkValueReader;
 import org.elasticsearch.hadoop.serialization.dto.mapping.FieldParser;
 import org.elasticsearch.hadoop.serialization.dto.mapping.MappingSet;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationErrorHandler;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationFailure;
-import org.elasticsearch.hadoop.serialization.handler.DeserializationHandlerLoader;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationErrorHandler;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationFailure;
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationHandlerLoader;
 import org.elasticsearch.hadoop.util.ObjectUtils;
 import org.elasticsearch.hadoop.util.TestSettings;
 import org.joda.time.format.ISODateTimeFormat;
